Tareas generales - Mostrar el tiempo de respuesta de una ejecucion - Mostrar la lista de pruebas con un nombre en vez del id - Mostrar los jsons en formato json y no en string - El menú de Bitácora lleva a la lista de ejecuciones realizadas

+++ src/main/java/com/quenti/smarttestui/service/dto/EjecucionPruebaDTO.java
@@ -16,7 +16,7 @@ public class EjecucionPruebaDTO implements Serializable {
 
     private LocalDate fecha;
 
-    private Integer tiempoRespuesta;
+    private Long tiempoRespuesta;
 
     private String resultado;
 
@@ -52,11 +52,11 @@ public class EjecucionPruebaDTO implements Serializable {
     public void setFecha(LocalDate fecha) {
         this.fecha = fecha;
     }
-    public Integer getTiempoRespuesta() {
+    public Long getTiempoRespuesta() {
         return tiempoRespuesta;
     }
 
-    public void setTiempoRespuesta(Integer tiempoRespuesta) {
+    public void setTiempoRespuesta(Long tiempoRespuesta) {
         this.tiempoRespuesta = tiempoRespuesta;
     }
     public String getResultado() {
